deploy/requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
dj-database-url==3.1.0
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
reportlab==4.4.9
sqlparse==0.5.5
tzdata==2025.3